fix: rename folder Navbar to navbar
